Removed the Threadexist test in the SetPositions() function to fix the problem where the position is set prior to the first run call. Re Remember the last media position in Zoomplayer fix.
Made some further improvments to the Load() file change feature to make it more robust and to allow for Infinite tee filter in the chain.
Fixed a number of bugs in the duration parser to allow for buggy streams such as those that have PRBS content. 

Made some further improvments to the Load() file change feature.
Some other minor improvements.


git-svn-id: http://svn.nate.deepcreek.org.au/svn/TSFileSource/trunk@73 89fdcf3f-49f1-7541-bbf8-3596d8354d9d