Notes taken from the forum
packet injection is now optional, as well as the use of the ParserSink(Shared Memory) instead of the FileSink filter for the internal network source graph. The Options are available from the PropPage & ITSFileSource Interface. All settings are now moved to the HKEY_CURRENT_USER registry location. Also the AC3 Audio & Audio 2 Preference options now forces a audio stream change when changed.


git-svn-id: http://svn.nate.deepcreek.org.au/svn/TSFileSource/trunk@127 89fdcf3f-49f1-7541-bbf8-3596d8354d9d